Adds ability to set configuration from outside of the class.
Renamed the external function, added an optional second argument.
This argument is an array that accepts certain keys - if the keys match the supported list of
properties that can be set, then they are set.

All the strings in the class have been extracted into the $strings property.
Ideally, the sprintf arguments would not be split up from the string containing the placeholders,
but when setting the defaults (as in the theme-file.php example), variables like $plugin or $this->menu are not in scope.
Instead, comments have been added to the end of the example code, to show what each placeholder argument would be.