Simplify plugin registration arguments.
The plugin path (e.g. 'edit-howdy/edithowdy.php') has now been simplified down to just needing the plugin slug (e.g. 'edit-howdy'). The developer does not need to find out which is the active plugin file.
We loop through the installed plugins early on, and search the array keys to determine what the fill path is, and assign that
to the $plugin['file_path'].

If the source of the plugin is the latest stable version from the .org repo, then the source argument does not need to be included, since this is populated from the Plugin API lookup that is done.