Allow adding a plugin parameter named `is_callable`
Here's a use case: Yoast's WordPress SEO Plugin and the Premium version. Mostly same plugin but different slugs. This allows you to say "no, the slug doesn't exist, but this function (or method) is callable, so the plugin must be active".

Here's the plugin array item passed to `tgmpa()`: 

``
array(
    'name'      => 'WordPress SEO by Yoast',
    'slug'      => 'wordpress-seo',
    'required'  => false,
    'is_callable' => 'wpseo_init',
)
``

If the `wpseo_init()` function exists, the plugin's still considered active.