Bug fix: only the user which switched theme would be re-notified
There may be several admin users for a site. If a theme using TGMPA is deactivated, the dismiss notice state would only be removed for the user doing the theme switching, not for all users.
This effectively meant that only that user would be re-notified of open recommended/required plugins if the theme would be reactivated.

This fix ensures that all eligable users will be re-notified on reatcivation of the theme using TGMPA.