Lean loading: only really load TGMPA when on the admin side
Needs review - AFAICS everything hooks into admin side hooks, so there shouldn't be a need to even register plugins/config when not in admin.
The only caveat to this I can think of is in a develop environment where some developers use a 'switch_theme' plugin which can be used from the admin bar.