Rename a filter for consistency.
An already existing filter on individual items is called `tgmpa_table_data_item`. It would therefore be more logical for this filter on the array of items to be called `tgmpa_table_data_items` instead of `tgmpa_plugin_table_items`.

This shouldn't give any issues with "existing implementations" as the filter was only introduced about a week ago.