Add a Custom TGMPA Generator to the website.
The Custom TGMPA Generator takes the zip file from the latest release (for now: needs to be uploaded after each release as I can't find a hard link to the zip on GitHub) and adjusts the contents based on settings provided by the end-user in a form in the download page of the website and then serves the customized .zip file to the user.

The output will be different depending on whether TGMPA will be used within a parent/child-theme or a plugin and the chosen publication channel: wordpress.org, Themeforest or other.

Other changes made to the website to support this:
* removed most links to the GH downloads and redirected them to the download page to encourage the use of the form.
* added a number of javascript libraries used by the custom generator which will conditionally load - i.e. only on the download page.
* added the v2.5.2 release zip file to use as base for the generator.
* added a blogpost announcing the custom generator