Update for PHPCompatibilityWP 2.0.0
Update to use the newer version of [PHPCompatibilityWP as released today](https://github.com/PHPCompatibility/PHPCompatibilityWP/releases/tag/2.0.0).

Also: not sure why I didn't change the standard reference in the ruleset in 755, but that's fixed now (not that it really mattered as TGMPA doesn't use any of the WP provided backfills, but still).