Exclude example plugin from dist
Github uses `git archive` to produce the automatic dist archive. This archive is what is downloaded when Composer retrieves the dependency by default or with `prefer-dist`. By adding this line, the plugin archive is excluded from dist because it is an example that is never needed in dist environments (like production), reducing download size and time.